Moved `tablewizard` package to `wizard.table`.
