+++ resources/META-INF/plugin.xml
@@ -159,7 +159,7 @@
             <add-to-group group-id="EditMenu" anchor="last"/>
 
             <!-- Table Creation Wizard-->
-            <action class="nl.hannahsten.texifyidea.action.tablewizard.LatexTableWizardAction" id="texify.TableCreationWizzard" text="_Table Creation Wizard"
+            <action class="nl.hannahsten.texifyidea.action.wizard.table.LatexTableWizardAction" id="texify.TableCreationWizzard" text="_Table Creation Wizard"
                     description="Shows popup to create and insert a table."/>
 
             <!-- Toggle Star -->
