+++ src/nl/hannahsten/texifyidea/action/wizard/table/LatexTableWizardAction.kt
@@ -1,4 +1,4 @@
-package nl.hannahsten.texifyidea.action.tablewizard
+package nl.hannahsten.texifyidea.action.wizard.table
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -21,6 +21,7 @@ import java.util.*
  * @author Abby Berkers
  */
 class LatexTableWizardAction : AnAction() {
+
     override fun actionPerformed(e: AnActionEvent) {
         val file = e.getData(PlatformDataKeys.VIRTUAL_FILE) ?: return
         val project = e.getData(PlatformDataKeys.PROJECT)
@@ -44,11 +45,11 @@ class LatexTableWizardAction : AnAction() {
 
             // Insert the booktabs package.
             WriteCommandAction.runWriteCommandAction(
-                project,
-                "Insert table",
-                "LaTeX",
-                Runnable { file.psiFile(project)!!.insertUsepackage(LatexPackage.BOOKTABS) },
-                file.psiFile(project)
+                    project,
+                    "Insert table",
+                    "LaTeX",
+                    Runnable { file.psiFile(project)!!.insertUsepackage(LatexPackage.BOOKTABS) },
+                    file.psiFile(project)
             )
         }
     }
@@ -68,17 +69,17 @@ class LatexTableWizardAction : AnAction() {
         fun processTableContent(indent: String): String {
             return with(tableInformation) {
                 val headers = tableModel.getColumnNames()
-                    .joinToString(
-                        prefix = "$indent\\toprule\n$indent",
-                        separator = " & ",
-                        postfix = " \\\\\n$indent\\midrule\n"
-                    ) { "\\textbf{$it}" }
+                        .joinToString(
+                                prefix = "$indent\\toprule\n$indent",
+                                separator = " & ",
+                                postfix = " \\\\\n$indent\\midrule\n"
+                        ) { "\\textbf{$it}" }
 
                 val rows = tableModel.dataVector.joinToString(separator = "\n", postfix = "\n$indent\\bottomrule\n") { row ->
                     (row as Vector<*>).joinToString(
-                        prefix = indent,
-                        separator = " & ",
-                        postfix = " \\\\"
+                            prefix = indent,
+                            separator = " & ",
+                            postfix = " \\\\"
                     ) {
                         // Enclose with $ if the type of this column is math.
                         val index = row.indexOf(it)
@@ -94,18 +95,18 @@ class LatexTableWizardAction : AnAction() {
         // (this includes the caption and label).
         return with(tableInformation) {
             val openTableCommand = "\\begin{table}\n" +
-                "$lineIndent$tabIndent\\centering\n" +
-                // Everything within the table command gets an extra indent.
-                "$lineIndent$tabIndent\\begin{tabular}{${columnTypes.toLatexColumnFormatters()}}\n"
+                    "$lineIndent$tabIndent\\centering\n" +
+                    // Everything within the table command gets an extra indent.
+                    "$lineIndent$tabIndent\\begin{tabular}{${columnTypes.toLatexColumnFormatters()}}\n"
 
             // The content has to be indented once more.
             val content = processTableContent(indent = lineIndent + tabIndent + tabIndent)
 
             val closeTableCommand = "$lineIndent$tabIndent\\end{tabular}\n" +
-                "$lineIndent$tabIndent\\caption{$caption}\n" +
-                "$lineIndent$tabIndent\\label{$label}\n" +
-                "$lineIndent\\end{table}\n" +
-                lineIndent // Indentation on the last line so we can continue typing there.
+                    "$lineIndent$tabIndent\\caption{$caption}\n" +
+                    "$lineIndent$tabIndent\\label{$label}\n" +
+                    "$lineIndent\\end{table}\n" +
+                    lineIndent // Indentation on the last line so we can continue typing there.
 
             openTableCommand + content + closeTableCommand
         }
