+++ src/nl/hannahsten/texifyidea/action/wizard/table/TableInformation.kt
@@ -1,4 +1,4 @@
-package nl.hannahsten.texifyidea.action.tablewizard
+package nl.hannahsten.texifyidea.action.wizard.table
 
 import nl.hannahsten.texifyidea.ui.tablecreationdialog.ColumnType
 import nl.hannahsten.texifyidea.ui.tablecreationdialog.TableCreationTableModel
